serial: transfer API to header
