+++ include/sddf/serial/queue.h
@@ -48,7 +48,25 @@ typedef struct serial_queue_handle {
  * @param buffer_init 1 indicates the head and tail indices in shared memory need to be initialised.
  *                    0 inidicates they do not. Only one side of the shared memory regions needs to do this.
  */
-void serial_queue_init(serial_queue_handle_t *queue, serial_queue_t *free, serial_queue_t *active, int buffer_init, uint32_t free_size, uint32_t active_size);
+void serial_queue_init(serial_queue_handle_t *queue, serial_queue_t *free, serial_queue_t *active, int buffer_init, uint32_t free_size, uint32_t active_size)
+{
+    queue->free = free;
+    queue->active = active;
+    if (buffer_init) {
+        queue->free->tail = 0;
+        queue->free->head = 0;
+        queue->free->size = free_size;
+        queue->free->notify_writer = false;
+        queue->free->notify_reader = false;
+        queue->free->plugged = false;
+        queue->active->tail = 0;
+        queue->active->head = 0;
+        queue->active->size = active_size;
+        queue->active->notify_writer =false;
+        queue->active->notify_reader = false;
+        queue->active->plugged = false;
+    }
+}
 
 /**
  * Check if the queue is empty.
@@ -57,7 +75,10 @@ void serial_queue_init(serial_queue_handle_t *queue, serial_queue_t *free, seria
  *
  * @return true indicates the buffer is empty, false otherwise.
  */
-int serial_queue_empty(serial_queue_t *queue);
+static inline int serial_queue_empty(serial_queue_t *queue)
+{
+    return !((queue->tail - queue->head) % queue->size);
+}
 
 /**
  * Check if the queue is full
@@ -66,9 +87,17 @@ int serial_queue_empty(serial_queue_t *queue);
  *
  * @return true indicates the buffer is full, false otherwise.
  */
-int serial_queue_full(serial_queue_t *queue);
+static inline int serial_queue_full(serial_queue_t *queue)
+{
+    // assert((queue->tail - queue->head) >= 0);
+    return !((queue->tail - queue->head + 1) % queue->size);
+}
 
-uint32_t serial_queue_size(serial_queue_t *queue);
+static inline uint32_t serial_queue_size(serial_queue_t *queue)
+{
+    // assert((queue->tail - queue->head) >= 0);
+    return (queue->tail - queue->head);
+}
 
 /**
  * Enqueue an element to a queue
@@ -79,7 +108,21 @@ uint32_t serial_queue_size(serial_queue_t *queue);
  *
  * @return -1 when queue is empty, 0 on success.
  */
-int serial_enqueue(serial_queue_t *queue, uintptr_t buffer, unsigned int len);
+static inline int serial_enqueue(serial_queue_t *queue, uintptr_t buffer, unsigned int len)
+{
+    // assert(buffer != 0);
+    if (serial_queue_full(queue)) {
+        return -1;
+    }
+
+    queue->entries[queue->tail % queue->size].encoded_addr = buffer;
+    queue->entries[queue->tail % queue->size].len = len;
+
+    THREAD_MEMORY_RELEASE();
+    queue->tail++;
+
+    return 0;
+}
 
 /**
  * Dequeue an element to a queue.
@@ -90,7 +133,22 @@ int serial_enqueue(serial_queue_t *queue, uintptr_t buffer, unsigned int len);
  *
  * @return -1 when queue is empty, 0 on success.
  */
-int serial_dequeue(serial_queue_t *queue, uintptr_t *addr, unsigned int *len);
+static inline int serial_dequeue(serial_queue_t *queue, uintptr_t *addr, unsigned int *len)
+{
+    if (serial_queue_empty(queue)) {
+        return -1;
+    }
+
+    // assert(queue->entries[queue->head % queue->size].encoded_addr != 0);
+
+    *addr = queue->entries[queue->head % queue->size].encoded_addr;
+    *len = queue->entries[queue->head % queue->size].len;
+
+    THREAD_MEMORY_RELEASE();
+    queue->head++;
+
+    return 0;
+}
 
 /**
  * Enqueue an element into an free queue.
@@ -102,7 +160,11 @@ int serial_dequeue(serial_queue_t *queue, uintptr_t *addr, unsigned int *len);
  *
  * @return -1 when queue is full, 0 on success.
  */
-int serial_enqueue_free(serial_queue_handle_t *queue, uintptr_t addr, unsigned int len);
+static inline int serial_enqueue_free(serial_queue_handle_t *queue, uintptr_t addr, unsigned int len)
+{
+    // assert(addr);
+    return serial_enqueue(queue->free, addr, len);
+}
 
 /**
  * Enqueue an element into a active queue.
@@ -114,7 +176,11 @@ int serial_enqueue_free(serial_queue_handle_t *queue, uintptr_t addr, unsigned i
  *
  * @return -1 when queue is full, 0 on success.
  */
-int serial_enqueue_active(serial_queue_handle_t *queue, uintptr_t addr, unsigned int len);
+static inline int serial_enqueue_active(serial_queue_handle_t *queue, uintptr_t addr, unsigned int len)
+{
+    // assert(addr);
+    return serial_enqueue(queue->active, addr, len);
+}
 
 /**
  * Dequeue an element from the free queue.
@@ -125,7 +191,10 @@ int serial_enqueue_active(serial_queue_handle_t *queue, uintptr_t addr, unsigned
  *
  * @return -1 when queue is empty, 0 on success.
  */
-int serial_dequeue_free(serial_queue_handle_t *queue, uintptr_t *addr, unsigned int *len);
+static inline int serial_dequeue_free(serial_queue_handle_t *queue, uintptr_t *addr, unsigned int *len)
+{
+    return serial_dequeue(queue->free, addr, len);
+}
 
 /**
  * Dequeue an element from a active queue.
@@ -136,21 +205,30 @@ int serial_dequeue_free(serial_queue_handle_t *queue, uintptr_t *addr, unsigned 
  *
  * @return -1 when queue is empty, 0 on success.
  */
-int serial_dequeue_active(serial_queue_handle_t *queue, uintptr_t *addr, unsigned int *len);
+static inline int serial_dequeue_active(serial_queue_handle_t *queue, uintptr_t *addr, unsigned int *len)
+{
+    return serial_dequeue(queue->active, addr, len);
+}
 
 /**
  * Set the plug of a queue to true.
  *
  * @param queue handle to plug.
 */
-void serial_queue_plug(serial_queue_t *queue);
+static inline void serial_queue_plug(serial_queue_t *queue)
+{
+    queue->plugged = true;
+}
 
 /**
  * Set the plug of a queue to false.
  *
  * @param queue handle to unplug.
 */
-void serial_queue_unplug(serial_queue_t *queue);
+static inline void serial_queue_unplug(serial_queue_t *queue)
+{
+    queue->plugged = false;
+}
 
 /**
  * Check the current value of the plug.
@@ -159,4 +237,7 @@ void serial_queue_unplug(serial_queue_t *queue);
  *
  * @return true when queue is plugged, false when unplugged.
 */
-bool serial_queue_plugged(serial_queue_t *queue);
+static inline bool serial_queue_plugged(serial_queue_t *queue)
+{
+    return queue->plugged;
+}
